Mehr nach MVVM, Trainerliste und Preisliste bug behoben.
