+++ Beatz_Fitness_App/Model.xcdatamodeld/Model.xcdatamodel/contents
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="21754" systemVersion="22E261" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
     <entity name="AusgefuehrterSatz" representedClassName="AusgefuehrterSatz" syncable="YES">
+        <attribute name="datum" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
         <attribute name="gewicht" optional="YES" attributeType="Double" defaultValueString="0.0" usesScalarValueType="YES"/>
         <attribute name="id" optional="YES" attributeType="UUID" usesScalarValueType="NO"/>
         <attribute name="uebungname" optional="YES" attributeType="String"/>
