+++ Beatz_Fitness_App/Split/ManuelleKlassen/AusgefuehrterSatz+CoreDataClass.swift
@@ -2,7 +2,7 @@
 //  AusgefuehrterSatz+CoreDataClass.swift
 //  Beatz_Fitness_App
 //
-//  Created by Maximillian Stabe on 30.04.23.
+//  Created by Maximillian Stabe on 14.05.23.
 //
 //
