+++ Beatz_Fitness_App/Split/ManuelleKlassen/AusgefuehrterSatz+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  AusgefuehrterSatz+CoreDataProperties.swift
 //  Beatz_Fitness_App
 //
-//  Created by Maximillian Stabe on 30.04.23.
+//  Created by Maximillian Stabe on 14.05.23.
 //
 //
 
@@ -16,13 +16,14 @@ extension AusgefuehrterSatz {
         return NSFetchRequest<AusgefuehrterSatz>(entityName: "AusgefuehrterSatz")
     }
 
-    @NSManaged public var id: UUID?
     @NSManaged public var gewicht: Double
-    @NSManaged public var wiederholungen: Int64
+    @NSManaged public var id: UUID?
     @NSManaged public var uebungname: String?
+    @NSManaged public var wiederholungen: Int64
+    @NSManaged public var datum: Date?
     @NSManaged public var saetze: NSOrderedSet?
     @NSManaged public var trainingseintrag: NSSet?
-    @NSManaged public var uebung: Uebung?
+    @NSManaged public var uebung: NSOrderedSet?
 
 }
 
@@ -75,6 +76,41 @@ extension AusgefuehrterSatz {
 
     @objc(removeTrainingseintrag:)
     @NSManaged public func removeFromTrainingseintrag(_ values: NSSet)
+
+}
+
+// MARK: Generated accessors for uebung
+extension AusgefuehrterSatz {
+
+    @objc(insertObject:inUebungAtIndex:)
+    @NSManaged public func insertIntoUebung(_ value: Uebung, at idx: Int)
+
+    @objc(removeObjectFromUebungAtIndex:)
+    @NSManaged public func removeFromUebung(at idx: Int)
+
+    @objc(insertUebung:atIndexes:)
+    @NSManaged public func insertIntoUebung(_ values: [Uebung], at indexes: NSIndexSet)
+
+    @objc(removeUebungAtIndexes:)
+    @NSManaged public func removeFromUebung(at indexes: NSIndexSet)
+
+    @objc(replaceObjectInUebungAtIndex:withObject:)
+    @NSManaged public func replaceUebung(at idx: Int, with value: Uebung)
+
+    @objc(replaceUebungAtIndexes:withUebung:)
+    @NSManaged public func replaceUebung(at indexes: NSIndexSet, with values: [Uebung])
+
+    @objc(addUebungObject:)
+    @NSManaged public func addToUebung(_ value: Uebung)
+
+    @objc(removeUebungObject:)
+    @NSManaged public func removeFromUebung(_ value: Uebung)
+
+    @objc(addUebung:)
+    @NSManaged public func addToUebung(_ values: NSOrderedSet)
+
+    @objc(removeUebung:)
+    @NSManaged public func removeFromUebung(_ values: NSOrderedSet)
 
 }
