--- Beatz_Fitness_App/Split/UebungsItemVM.swift
@@ -1,18 +1,0 @@
-//
-//  VordefinierteUebungen.swift
-//  Beatz_Fitness_App
-//
-//  Created by Maximillian Stabe on 10.04.23.
-//
-
-import SwiftUI
-
-struct UebungsItem: Identifiable {
-    var id = UUID()
-    let uebungName: String
-    let uebungBeschreibung: String
-    var anzahlSaetze: Int
-    let bild: Image
-}
-
-
